separated out ClickToCopy component
