+++ client/src/containers/AssetInfo/view.jsx
@@ -5,38 +5,12 @@ import RowLabeled from '@components/RowLabeled';
 import Row from '@components/Row';
 import SpaceBetween from '@components/SpaceBetween';
 import AssetShareButtons from '@components/AssetShareButtons';
-
-const ClickToCopy = ({id, value, copyToClipboard}) => {
-  return (
-    <input
-      id={id}
-      value={value}
-      onClick={copyToClipboard}
-      type='text'
-      className='click-to-copy'
-      readOnly
-      spellCheck='false'
-    />
-  );
-};
+import ClickToCopy from '@components/ClickToCopy';
 
 class AssetInfo extends React.Component {
   constructor (props) {
     super(props);
     this.copyToClipboard = this.copyToClipboard.bind(this);
-  }
-  copyToClipboard (event) {
-    console.log('event:', event);
-    console.log('event.target:', event.target);
-    console.log('event.target.id:', event.target.id);
-    const elementToCopy = event.target.id;
-    const element = document.getElementById(elementToCopy);
-    element.select();
-    try {
-      document.execCommand('copy');
-    } catch (err) {
-      this.setState({error: 'Oops, unable to copy'});
-    }
   }
   render () {
     const { asset: { shortId, claimData : { channelName, certificateId, description, name, claimId, fileExt, contentType, thumbnail, host } } } = this.props;
@@ -81,7 +55,6 @@ class AssetInfo extends React.Component {
               <ClickToCopy
                 id={'short-link'}
                 value={`${host}/${shortId}/${name}.${fileExt}`}
-                copyToClipboard={this.copyToClipboard}
               />
             }
           />
@@ -98,13 +71,11 @@ class AssetInfo extends React.Component {
                   <ClickToCopy
                     id={'embed-text-video'}
                     value={`<video width="100%" controls poster="${thumbnail}" src="${host}/${claimId}/${name}.${fileExt}"/></video>`}
-                    copyToClipboard={this.copyToClipboard}
                   />
                 ) : (
                   <ClickToCopy
                     id={'embed-text-image'}
                     value={`<img src="${host}/${claimId}/${name}.${fileExt}"/>`}
-                    copyToClipboard={this.copyToClipboard}
                   />
                 )}
               </div>
